apis/events: fix return value of `MapEvent`
This commit fixes two errors in the event returned by the
`(*collectedEvent).MapEvent` method.

+++ apis/events/collector.go
@@ -599,7 +599,7 @@ func mergeContexts(ctx, defaultCtx *eventContext) {
 func (c *collector) enrichEvent(event *collectedEvent) {
 
 	// Source.
-	event.source = int32(event.header.source)
+	event.source = event.header.source
 
 	// AnonymousId.
 	if event.AnonymousId == "" {
